backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.3.0
insightface>=0.7.3
onnxruntime>=1.16.0
numpy>=1.26.0
opencv-python-headless>=4.9.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.1.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
pandas>=2.1.0
openpyxl>=3.1.0
